# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HRPF API",
    "version": "0.4.0",
    "description": "Backend foundations, managed public content and atomic publication controls. Fixed NGO page copy is bundled in the frontend and needs no publication approval. Operational member/case review workflows remain later milestones. Public reads are no-store and withdrawals take effect immediately. Browser clients use the frontend same-origin /api rewrite and credentials. All mutating auth/admin requests require the exact configured Origin and X-CSRF-Token. Production cookies use Secure, HttpOnly, SameSite=Lax, Path=/, no Domain; development names omit __Host-."
  },
  "paths": {
...
    "/api/admin/publication/{kind}": {
      "get": {
        "summary": "Read bounded publication review records with versions; requires entity permission",
        "security": [
          {
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "blog",
                "project",
                "board",
                "gallery",
                "report",
                "certificate",
                "setting"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "propertyNames": {
                          "type": "string"
                        },
                        "additionalProperties": {}
                      }
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/publication/{kind}/{id}": {
      "post": {
        "summary": "Publish or withdraw a reviewed record atomically; version, CSRF, current role and audited release attestation required",
        "security": [
          {
            "csrfHeader": [],
            "csrfCookie": [],
            "accessCookie": []
          }
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "blog",
                "project",
                "board",
                "gallery",
                "report",
                "certificate",
                "setting"
              ]
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "published",
                            "withdrawn"
                          ]
                        },
                        "version": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "status",
                        "version"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Redacted validation, authorization, conflict, limit or dependency error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PublicationInput": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "action": {
            "type": "string",
            "enum": [
              "publish",
              "withdraw"
            ]
          },
          "releaseReviewed": {
            "type": "boolean",
            "const": true
          },
          "assetId": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{24}$"
          }
        },
        "required": [
          "version",
          "action",
          "releaseReviewed"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "code",
              "message",
              "requestId"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    },
    "securitySchemes": {
      "accessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-access"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-refresh"
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-hrpf-csrf"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "formTicketHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Form-Ticket"
      }
    }
  }
}